This has been the wrong path to go down
