+++ WMS/WMS.Common.DataContext.Sqlite/WMSContext.cs
@@ -1,68 +1,88 @@
-using Microsoft.EntityFrameworkCore;
-using WMS.Common.EntityModels.Sqlite;
-using Microsoft.EntityFrameworkCore.Metadata;
+﻿using Microsoft.EntityFrameworkCore;
 
-namespace WMS.Common.DataContext.Sqlite
+namespace WMS.Shared;
+
+public partial class WmsContext : DbContext
 {
-    public partial class WMSContext : DbContext
+    public WmsContext()
     {
-        public WMSContext()
-        { }
-
-        public WMSContext(DbContextOptions<WMSContext> options) : base(options)
-        {
-
-        }
-
-        // TODO DbSets, OnConfiguring and OnModelCreating 
-        public virtual DbSet<Employee> Employees { get; set; } = null!;
-        public virtual DbSet<Request> Requests { get; set; } = null!;
-        public virtual DbSet<Category> Categories { get; set; } = null!;
-        public virtual DbSet<Project> Projects { get; set; } = null!;
-        public virtual DbSet<Change> Changes { get; set; } = null!;
-        public virtual DbSet<Priority> Priorities { get; set; } = null!;
-        public virtual DbSet<Status> Statuses { get; set; } = null!;
-        public virtual DbSet<ApprovalStatus> ApprovalStatuses { get; set; } = null!;
-        public virtual DbSet<RequestType> RequestTypes { get; set; } = null!;
-        public virtual DbSet<Role> Roles { get; set; } = null!;
-        public virtual DbSet<Work> Works { get; set; } = null!;
-        public virtual DbSet<WorkItem> WorkItems { get; set; } = null!;
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                string dir = Environment.CurrentDirectory;
-                string path = string.Empty;
-
-                if (dir.EndsWith("net8.0"))
-                {
-                    path = Path.Combine("..", "..", "..", "..", "WMS.db");
-                }
-                else
-                {
-                    path = Path.Combine("..", "WMS.db");
-                }
-
-                optionsBuilder.UseSqlite($"Filename={path}");
-            }
-        }
-
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            // ! Requests
-            builder.Entity<Request>(entity =>
-            {
-                entity.HasKey(e => new { e.StatusId, e.RequestTypeId, e.ApprovalStatusId });
-
-                entity.HasMany(d => d.Requestors);
-            });
-
-            //builder.Entity
-        }
-
-        partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
     }
 
+    public WmsContext(DbContextOptions<WmsContext> options)
+        : base(options)
+    {
+    }
 
+    public virtual DbSet<ApprovalStatus> ApprovalStatuses { get; set; }
+
+    public virtual DbSet<Category> Categories { get; set; }
+
+    public virtual DbSet<Change> Changes { get; set; }
+
+    public virtual DbSet<Employee> Employees { get; set; }
+
+    public virtual DbSet<Priority> Priorities { get; set; }
+
+    public virtual DbSet<Statuses> Statuses { get; set; }
+
+    public virtual DbSet<Project> Projects { get; set; }
+
+    public virtual DbSet<Request> Requests { get; set; }
+
+    public virtual DbSet<RequestType> RequestTypes { get; set; }
+
+    public virtual DbSet<Role> Roles { get; set; }
+
+    public virtual DbSet<Work> Works { get; set; }
+
+
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        => optionsBuilder.UseSqlite("Filename=../WMS.db");
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<Change>(entity =>
+        {
+            entity.HasKey(e => new { e.ChangeId });
+            entity.HasOne(e => e.RequestNavigation)
+                .WithMany(e => e.Changes)
+                .HasForeignKey(e => e.Request);
+            entity.HasOne(e => e.RequestorNavigation)
+                .WithMany(e => e.Changes)
+                .HasForeignKey(e => e.Requestor);
+        });
+
+        // modelBuilder.Entity<Employee>(entity =>
+        // {
+        //     entity.HasKey(e => new { e.EmployeeId });
+        //     entity.HasMany(e => e.Roles)
+        //         .WithMany(e => e.Employees);
+        // });
+
+
+        modelBuilder.Entity<Project>(entity =>
+        {
+            entity.HasKey(e => new { e.ProjectId });
+            entity.HasOne(e => e.RequestNavigation)
+                .WithMany(e => e.Projects)
+                .HasForeignKey(e => e.Request);
+            entity.HasOne(e => e.RequestorNavigation)
+                .WithMany(e => e.Projects)
+                .HasForeignKey(e => e.Requestor);
+        });
+
+        modelBuilder.Entity<Request>(entity =>
+        {
+            entity.HasKey(e => new { e.RequestId });
+            entity.HasOne(e => e.Employee)
+                .WithMany(e => e.Requests)
+                .HasForeignKey(e => e.EmployeeId);
+        });
+
+
+        OnModelCreatingPartial(modelBuilder);
+    }
+
+    partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
 }
