+++ WMS/WMS.Common.EntityModels.Sqlite/Statuses.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,6 +9,12 @@ namespace WMS.Shared
 {
     public class Statuses
     {
+        [Key]
+        [Column(TypeName = "GUID")]
+        public Guid StatusId { get; set; }
 
+        [Column(TypeName = "nvarchar(25)")]
+        [StringLength(25)]
+        public string StatusName { get; set; } = null!;
     }
 }
