+++ WMS/WMS.Common.EntityModels.Sqlite/scripts/WMS4SQLite.sql
@@ -13,7 +13,8 @@ DROP TABLE IF EXISTS "Works";
 DROP TABLE IF EXISTS "WorkItems";
 
 -- Employee Table
-CREATE TABLE "Employees" (
+CREATE TABLE "Employees"
+(
     "EmployeeId" GUID PRIMARY KEY,
     "LastName" nvarchar (20) NOT NULL,
     "FirstName" nvarchar (20) NOT NULL,
@@ -22,19 +23,22 @@ CREATE TABLE "Employees" (
     "PhoneNumber" nvarchar (24) NULL,
     "Extension" nvarchar (10) NULL,
     "ReportsTo" GUID NULL,
-    "Role" GUID NULL,
+    "RoleId" GUID NULL,
     "Notes" ntext NULL,
     CONSTRAINT "FK_Employees_Roles" FOREIGN KEY
     (
-        "Role"
+        "RoleId"
     ) REFERENCES "Roles" (
         "RoleId"
     )
 );
 CREATE INDEX "LastName" ON "Employees"("LastName");
+CREATE INDEX "RolesEmployee" ON "Employees"("RoleId");
+CREATE INDEX "RoleId" ON "Employees"("RoleId");
 
 -- Categories Table
-CREATE TABLE "Categories" (
+CREATE TABLE "Categories"
+(
     "CategoryId" GUID PRIMARY KEY,
     "CategoryName" nvarchar (15) NOT NULL,
     "Description" ntext NULL,
@@ -44,7 +48,8 @@ CREATE INDEX "CategoryName" ON "Categories"("CategoryName");
 
 -- ApprovalStatuses Table
 -- Stand alone table to store set values (pending, approved, denied)
-CREATE TABLE "ApprovalStatuses" (
+CREATE TABLE "ApprovalStatuses"
+(
     "AprovalStatusId" GUID PRIMARY KEY,
     "ApprovalStatusName" nvarchar(25) NOT NULL,
 );
@@ -52,7 +57,8 @@ CREATE TABLE "ApprovalStatuses" (
 
 -- Priorities Table
 -- Stand alone table to store set values (low, standard, high, emergency)
-CREATE TABLE "Priorities" (
+CREATE TABLE "Priorities"
+(
     "PriorityId" GUID PRIMARY KEY,
     "PriorityName" nvarchar (20) NOT NULL,
 );
@@ -60,7 +66,8 @@ CREATE TABLE "Priorities" (
 
 --Request Type
 -- Stand alone table to store set values (change request, project request)
-CREATE TABLE "RequestTypes" (
+CREATE TABLE "RequestTypes"
+(
     "RequestTypeId" GUID PRIMARY KEY,
     "RequestTypeName" NVARCHAR (25) NOT NULL,
 );
@@ -68,22 +75,25 @@ CREATE TABLE "RequestTypes" (
 --Roles 
 --Table to store set values, can be used many times to a record
 --(staff, change manager, project manager, board member, admin)
-CREATE TABLE  "Roles" (
+CREATE TABLE "Roles"
+(
     "RoleId" GUID PRIMARY KEY,
     "RoleName" NVARCHAR (25) NOT NULL,
 );
-CREATE INDEX "EmployeeRole" ON "Roles"("RoleName");
+CREATE INDEX "EmployeeRole" ON "Roles"("RoleId");
 
 --Status
 -- Stand alone table to store set values (in progess, on hold, cancelled, completed, pending)
-CREATE TABLE "Statuses" (
+CREATE TABLE "Statuses"
+(
     "StatusId" GUID PRIMARY KEY,
     "StatusName" NVARCHAR (25) NOT NULL,
 );
 
 --Request
 --Table for a request submitted by an employee
-CREATE TABLE "Requests" (
+CREATE TABLE "Requests"
+(
     "RequestId" GUID PRIMARY KEY,
     "RequestTypeId" GUID NULL,
     "StatusId" GUID NULL,
@@ -120,7 +130,8 @@ CREATE TABLE "Requests" (
 CREATE INDEX "EmployeeId" ON "Requests"("EmployeeId");
 CREATE INDEX "EmployeesRequests" ON "Requests"("EmployeeId");
 
-CREATE TABLE "Projects" (
+CREATE TABLE "Projects"
+(
     "ProjectId" GUID PRIMARY KEY,
     "StatusId" GUID NULL,
     "ApprovalStatusId" GUID NULL,
@@ -130,7 +141,7 @@ CREATE TABLE "Projects" (
     "PriorityId" GUID NULL,
     "Work" GUID NULL,
     "ProjectName" NVARCHAR (50) NOT NULL,
-    "Description" NTEXT NULL, 
+    "Description" NTEXT NULL,
     CONSTRAINT "FK_Projects_Status" FOREIGN KEY
     (
         "StatusId"
@@ -180,7 +191,8 @@ CREATE INDEX "EmployeesProjects" ON "Projects"("Requestor");
 CREATE INDEX "RequestsProjects" ON "Projects"("Request");
 CREATE INDEX "WorksProjects" ON "Projects"("Work");
 
-CREATE TABLE "Changes" (
+CREATE TABLE "Changes"
+(
     "ChangeId" GUID PRIMARY KEY,
     "PriorityId" GUID NULL,
     "StatusId" GUID NULL,
@@ -240,7 +252,8 @@ CREATE INDEX "EmployeesChanges" ON "Changes"("Requestor");
 CREATE INDEX "RequestsChanges" ON "Changes"("Request");
 CREATE INDEX "WorksChanges" ON "Changes"("Work");
 
-CREATE TABLE "Works" (
+CREATE TABLE "Works"
+(
     "WorkId" GUID PRIMARY KEY
     "Project" GUID NULL,
     "Change" GUID NULL,
@@ -269,7 +282,8 @@ CREATE INDEX "WorksProjects" ON "Works"("Project");
 CREATE INDEX "WorksChanges" ON "Works"("Change");
 CREATE INDEX "WorksWorkItems" ON "Works"("WorkItems");
 
-CREATE TABLE "WorkItems" (
+CREATE TABLE "WorkItems"
+(
     "WorkItemId" GUID PRIMARY KEY,
     "WorkId" GUID NULL,
     "PriorityId" GUID NULL,
